+++ src/com/lucienrowan/financemanager/objects/Months.java
@@ -1,0 +1,16 @@
+package com.lucienrowan.financemanager.objects;
+
+public enum Months {
+    JANUARY,
+    FEBRUARY,
+    MARCH,
+    APRIL,
+    MAY,
+    JUNE,
+    JULY,
+    AUGUST,
+    SEPTEMBER,
+    OCTOBER,
+    NOVEMBER,
+    DECEMBER
+}