+++ src/com/lucienrowan/financemanager/objects/TimeGroup.java
@@ -3,35 +3,29 @@ package com.lucienrowan.financemanager.objects;
 public class TimeGroup {
     private int id;
     private int year;
-    private int month;
+    private Months month;
     private int iteration;
 
-    public TimeGroup(int year, int month) {
+    public TimeGroup(int year, Months month) {
         this.year = year;
         setMonth(month);
     }
 
-    public TimeGroup(int year, int month, int iteration) {
+    public TimeGroup(int year, Months month, int iteration) {
         this.year = year;
         setMonth(month);
         setIteration(iteration);
     }
 
-    public TimeGroup(int id, int year, int month, int iteration) {
+    public TimeGroup(int id, int year, Months month, int iteration) {
         this.id = id;
         this.year = year;
         setMonth(month);
         setIteration(iteration);
     }
 
-    public void setMonth(int month) {
-        if (month >= 1 && month <= 12) {
-            this.month = month;
-        }
-
-        else {
-            throw new IllegalArgumentException("Invalid month value.");
-        }
+    public void setMonth(Months month) {
+        this.month = month;
     }
 
     public void setIteration(int iteration) {